reafctor: Refactors Avatar component and moves imageMocks and imageControl to DropDownItem story
